LVQ moved
is it needed?
